requirements.txt
asgiref==3.2.10
crispy-forms-materialize==0.2
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-environ==0.4.5
django-heroku==0.3.1
django-materialize==1.0.1a2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
